Don't log templates as components
Some templates are logging errors with componentData.name set to the
template name (eg: auto-update-history-on-this-day). This causes the
eventApp field to show [template-name] instead of `HTML Template:
${templateName}`.

This correction forces the "HTML Template: " prefix when the
component.name exactly matches the template name so that the eventApp
consistently shows the prefix for templates.